requirements.txt
aiosqlite
discord
python-dotenv
spotipy
